+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/provisioning/serviceinstance/FetchServiceInstanceProvider.groovy
@@ -1,0 +1,7 @@
+package com.swisscom.cloud.sb.broker.provisioning.serviceinstance
+
+import com.swisscom.cloud.sb.broker.model.ServiceInstance
+
+interface FetchServiceInstanceProvider {
+    ServiceInstanceResponseDto fetchServiceInstance(ServiceInstance instance)
+}