+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/servicedefinition/ServiceDefinitionProcessor.groovy
@@ -91,6 +91,9 @@ class ServiceDefinitionProcessor {
         service.serviceProviderClass = serviceJson.serviceProviderClass
         service.displayIndex = serviceJson.displayIndex
         service.asyncRequired = serviceJson.asyncRequired
+        service.plan_updateable = serviceJson.plan_updateable
+        service.instancesRetrievable = serviceJson.instancesRetrievable
+        service.bindingsRetrievable = serviceJson.bindingsRetrievable
         // dashboard items
         if (serviceJson.dashboard_client) {
             service.dashboardClientId = serviceJson.dashboard_client.id