+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/util/test/DummyServiceProviderParameters.groovy
@@ -1,0 +1,8 @@
+package com.swisscom.cloud.sb.broker.util.test
+
+class DummyServiceProviderParameters {
+    String mode
+    Integer delay
+    String success
+    String parentReference
+}